Fix ocassional unmarshalling error
Described in issue #3, this happens because Kibana API could return
floating point numbers for uptime.

This commit also fixes a bug where a trailing slash in the kibana.uri
argument could fail connecting to the API.

This change also adds an additional debug flag to output Kibana API
response to stdout for debugging purposes.

+++ exporter/collector.go
@@ -34,7 +34,7 @@ type KibanaMetrics struct {
 	Metrics struct {
 		ConcurrentConnections int `json:"concurrent_connections"`
 		Process               struct {
-			UptimeInMillis int64 `json:"uptime_in_millis"`
+			UptimeInMillis float64 `json:"uptime_in_millis"`
 			Memory         struct {
 				Heap struct {
 					TotalInBytes int64 `json:"total_in_bytes"`
